+++ apps/cowswap-frontend/src/modules/application/containers/App/index.tsx
@@ -1,11 +1,10 @@
-import { ReactNode, Suspense } from 'react'
+import { ReactNode } from 'react'
 
 import ErrorBoundary from 'legacy/components/ErrorBoundary'
 import { TopLevelModals } from 'legacy/components/TopLevelModals'
 
 import { InvalidCoWShedSetup } from 'modules/cowShed'
 
-import { LoadingApp } from 'common/pure/LoadingApp'
 import RedirectAnySwapAffectedUsers from 'pages/error/AnySwapAffectedUsers/RedirectAnySwapAffectedUsers'
 
 import { RoutesApp } from './RoutesApp'
@@ -15,15 +14,13 @@ import { AppContainer } from '../AppContainer'
 export function App(): ReactNode {
   return (
     <ErrorBoundary>
-      <Suspense fallback={<LoadingApp />}>
-        <RedirectAnySwapAffectedUsers />
-        <InvalidCoWShedSetup />
+      <RedirectAnySwapAffectedUsers />
+      <InvalidCoWShedSetup />
 
-        <AppContainer>
-          <TopLevelModals />
-          <RoutesApp />
-        </AppContainer>
-      </Suspense>
+      <AppContainer>
+        <TopLevelModals />
+        <RoutesApp />
+      </AppContainer>
     </ErrorBoundary>
   )
 }