+++ apps/cowswap-frontend/src/modules/swap/containers/SwapConfirmModal/index.tsx
@@ -139,6 +139,7 @@ export function SwapConfirmModal(props: SwapConfirmModalProps): ReactNode {
                   recipientChainId={bridgeContext?.buyAmount?.currency?.chainId}
                 />
                 {restContent}
+                <HighFeeWarning readonlyMode />
               </>
             )
           : (restContent) => (