+++ apps/cowswap-frontend/src/modules/tradeWidgetAddons/containers/HighFeeWarning/hooks/useHighFeeWarning.ts
@@ -1,7 +1,8 @@
 import { atom, useAtom } from 'jotai'
-import { useMemo } from 'react'
+import { SetStateAction, useMemo } from 'react'
 
 import { FEE_SIZE_THRESHOLD } from '@cowprotocol/common-const'
+import { Fraction } from '@uniswap/sdk-core'
 
 import { useGetReceiveAmountInfo } from 'modules/trade'
 
@@ -10,27 +11,36 @@ import { useSafeEffect, useSafeMemo } from 'common/hooks/useSafeMemo'
 const feeWarningAcceptedAtom = atom(false)
 const DEFAULT_FEE_STATE: [boolean, undefined] = [false, undefined]
 
+type SetAtom<Args extends unknown[], Result> = (...args: Args) => Result
+
+interface UseHighFeeWarningReturn {
+  isHighFee: boolean
+  feePercentage: Fraction | undefined
+  isHighBridgeFee: boolean | undefined
+  bridgeFeePercentage: Fraction | undefined
+  feeWarningAccepted: boolean
+  setFeeWarningAccepted: SetAtom<[SetStateAction<boolean>], void>
+}
+
 /**
  * useHighFeeWarning
  * @description checks whether fee vs trade inputAmount = high fee warning
  * @description returns params related to high fee and a cb for checking/unchecking fee acceptance
  */
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useHighFeeWarning() {
+export function useHighFeeWarning(): UseHighFeeWarningReturn {
   const receiveAmountInfo = useGetReceiveAmountInfo()
 
   const [feeWarningAccepted, setFeeWarningAccepted] = useAtom(feeWarningAcceptedAtom)
 
   // only considers inputAmount vs fee (fee is in input token)
-  const [isHighFee, feePercentage] = useMemo(() => {
+  const [isHighFee, feePercentage, isHighBridgeFee, bridgeFeePercentage] = useMemo(() => {
     if (!receiveAmountInfo) return DEFAULT_FEE_STATE
 
     const {
       isSell,
       beforeNetworkCosts,
       afterNetworkCosts,
-      costs: { networkFee, partnerFee },
+      costs: { networkFee, partnerFee, bridgeFee },
       quotePrice,
     } = receiveAmountInfo
 
@@ -49,7 +59,11 @@ export function useHighFeeWarning() {
 
     const feePercentage = totalFeeAmount.divide(targetAmount).multiply(100).asFraction
 
-    return [feePercentage.greaterThan(FEE_SIZE_THRESHOLD), feePercentage]
+    const bridgeFeePercentage = bridgeFee?.amountInDestinationCurrency.divide(targetAmount).multiply(100).asFraction
+
+    const isHighBridgeFee = Boolean(bridgeFeePercentage?.greaterThan(FEE_SIZE_THRESHOLD))
+
+    return [feePercentage.greaterThan(FEE_SIZE_THRESHOLD), feePercentage, isHighBridgeFee, bridgeFeePercentage]
   }, [receiveAmountInfo])
 
   // reset the state when users change swap params
@@ -65,27 +79,29 @@ export function useHighFeeWarning() {
     () => ({
       isHighFee,
       feePercentage,
+      isHighBridgeFee,
+      bridgeFeePercentage,
       // we only care/check about feeWarning being accepted if the fee is actually high..
-      feeWarningAccepted: _computeFeeWarningAcceptedState({ feeWarningAccepted, isHighFee }),
+      feeWarningAccepted: _computeFeeWarningAcceptedState({ feeWarningAccepted, isHighFee, isHighBridgeFee }),
       setFeeWarningAccepted,
     }),
-    [isHighFee, feePercentage, feeWarningAccepted, setFeeWarningAccepted],
+    [isHighFee, feePercentage, isHighBridgeFee, bridgeFeePercentage, feeWarningAccepted, setFeeWarningAccepted],
   )
 }
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 function _computeFeeWarningAcceptedState({
   feeWarningAccepted,
   isHighFee,
+  isHighBridgeFee,
 }: {
   feeWarningAccepted: boolean
   isHighFee: boolean
-}) {
+  isHighBridgeFee: boolean | undefined
+}): boolean {
   if (feeWarningAccepted) return true
   else {
     // is the fee high? that's only when we care
-    if (isHighFee) {
+    if (isHighFee || isHighBridgeFee) {
       return feeWarningAccepted
     } else {
       return true